02-functions/code/delete/requirements.txt
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.10.0
